Add ability to exclude paths explicitly from syncing
This patch expands the syntax of the 'sync_list' file. If a line is
prefixed with '-' the rest of the line is a path to be excluded from
syncing. If it is prefixed with '+' the rest of the line is a path to be
synced.

This allows to sync a directory, but to exclude a sub directory of it,