This patch allows to specify a driveId in the config file via the 'drive_id' parameter. If this parameter is given, the driveUrl is modified so all requests will go to the specified driveId in stead of the users drive.
The file README.odrive.md shows, how to get the driveId of a Sharepoint
group.

We removed unused code which could create problems with future additions.

The variable itemByIdUrl is prefixed for the personal OneDrive drive.
However it is never used in the code. The itemByIdUrl is constructed on
the fly where needed.

The variable itemByPathUrl ditto, however is used in viewChangesByPath(),
a function which in turn is never used in the code.  The function seems
also to calculate the resulting path url incorrectly.

All three are removed, so if somebody wants to implement itemByPath
functionality in the future she does not stumble over hardcoded paths
which would have to be fixed if used on a none personal drive.